% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTreatmentSequence-package.R
\docType{package}
\name{ggTreatmentSequence-package}
\alias{ggTreatmentSequence}
\alias{ggTreatmentSequence-package}
\title{ggTreatmentSequence: Creates Treatment Pattern Summaries}
\description{
Creates Treatment Pattern Summaries
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mattsecrest.github.io/ggTreatmentSequence/}
}

}
\author{
\strong{Maintainer}: Matthew Secrest \email{secrmatt@gmail.com} (\href{https://orcid.org/0000-0002-0939-4902}{ORCID})

}
\keyword{internal}
