% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtxseq.R
\name{ggtxseq}
\alias{ggtxseq}
\title{Sankey Plot of Treatment Sequence}
\usage{
ggtxseq()
}
\description{
Function to create a "sankey plot" aka alluvial plot of treatment sequences
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
